+++ src/views/CodeCoverageTreeViewProvider.ts
@@ -1,5 +1,7 @@
 import * as vscode from 'vscode';
 import { ApexClass } from '../classes/Apex';
+import { PinnedClasses } from '../common/PinnedClasses';
+import { usePinnedClassIcon } from './pinnedClassTreeItem';
 
 export class CodeCoverageTreeViewProvider implements vscode.TreeDataProvider<vscode.TreeItem> {
   private _onDidChangeTreeData: vscode.EventEmitter<vscode.TreeItem | undefined | void> =
@@ -8,6 +10,7 @@ export class CodeCoverageTreeViewProvider implements vscode.TreeDataProvider<vsc
     this._onDidChangeTreeData.event;
 
   private _apexClasses: ApexClass[] | undefined = undefined;
+  private readonly pinnedClasses: PinnedClasses;
 
   get apexClasses(): ApexClass[] | undefined {
     return this._apexClasses;
@@ -21,7 +24,8 @@ export class CodeCoverageTreeViewProvider implements vscode.TreeDataProvider<vsc
     }
   }
 
-  constructor() {
+  constructor(pinnedClasses = new PinnedClasses()) {
+    this.pinnedClasses = pinnedClasses;
     this.apexClasses = undefined;
   }
 
@@ -53,8 +57,12 @@ export class CodeCoverageTreeViewProvider implements vscode.TreeDataProvider<vsc
       return children;
     }
 
-    this.apexClasses.map((apexClass) => {
-      children.push(apexClass.getTreeItem());
+    this.pinnedClasses.order('codeCoverage', this.apexClasses).map((apexClass) => {
+      const item = apexClass.getTreeItem();
+      const isPinned = this.pinnedClasses.isPinned('codeCoverage', apexClass.name);
+      item.contextValue = isPinned ? 'pinnedApexCoverageClass' : 'apexCoverageClass';
+      if (isPinned) usePinnedClassIcon(item);
+      children.push(item);
     });
 
     return children;